feat: clearer message

+++ aksa-be/app/Http/Controllers/AuthenticatedSessionController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreSessionRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\ValidationException;
@@ -9,14 +10,11 @@ use App\Http\Resources\UserResource;
 
 class AuthenticatedSessionController extends Controller
 {
-    public function store(Request $request)
+    public function store(StoreSessionRequest $request)
     {
-        $request->validate([
-            'username' => ['required', 'string'],
-            'password' => ['required', 'string'],
-        ]);
+        $validated = $request->validated();
 
-        if (!Auth::attempt($request->only('username', 'password'))) {
+        if (!Auth::attempt($validated)) {
             throw ValidationException::withMessages([
                 'username' => trans('auth.failed'),
             ]);
@@ -26,15 +24,26 @@ class AuthenticatedSessionController extends Controller
 
         $token = $user->createToken('auth_token')->plainTextToken;
 
-        return $this->successResponse([
-            $user->role => new UserResource($user),
-            'token' => $token,
-        ], 200);
+        return $this->successResponse(
+            data: [
+                $user->role => new UserResource($user),
+                'token' => $token,
+            ],
+            message: 'Login successfully.',
+            code: 200
+        );
     }
 
     public function destroy(Request $request)
     {
+        if (!Auth::check()) {
+            return $this->errorResponse(message: 'No active session to logout from.', code: 401);
+        }
+
         $request->user()->currentAccessToken()->delete();
-        return $this->successResponse(message: 'Logout successfully.');
+        return $this->successResponse(
+            message: 'Logout successfully.',
+            code: 204
+        );
     }
 }
